+++ RedPandaIDE/settingsdialog/projectfileswidget.h
@@ -63,6 +63,10 @@ private slots:
 
 public:
     void init() override;
+
+    // QWidget interface
+protected:
+    void showEvent(QShowEvent *event) override;
 };
 
 #endif // PROJECTFILESWIDGET_H